Added Inital Setup Of navigation
