+++ app/src/main/java/com/nervesparks/iris/ui/ParametersScreen.kt
@@ -1,0 +1,10 @@
+package com.nervesparks.iris.ui
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+
+fun ParametersScreen(){
+    Text(text = "This is Parameters Screen")
+}
