+++ app/src/main/java/com/nervesparks/iris/ui/SearchResultScreen.kt
@@ -1,0 +1,9 @@
+package com.nervesparks.iris.ui
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SearchResultScreen(){
+    Text(text = "This is Search Screen")
+}
